Zoom Feature & Crouch Edit
I now made it so that when the RMB is clicked, the player will zoom in. When released, the player camera will return to the original position.

Also made a few choice edits of the crouch script; it will now return the player to their original size when entering the Air or Aiming states